query fuzzing: test cleaning and renaming
