+++ go/test/endtoend/vtgate/queries/query_fuzzing/dept_insert.sql
@@ -1,0 +1,5 @@
+INSERT INTO dept(deptno, dname, loc)
+VALUES ('10', 'ACCOUNTING', 'NEW YORK'),
+       ('20', 'RESEARCH', 'DALLAS'),
+       ('30', 'SALES', 'CHICAGO'),
+       ('40', 'OPERATIONS', 'BOSTON')
