+++ go/test/endtoend/vtgate/queries/query_fuzzing/query_gen_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package random
+package query_fuzzing
 
 import (
 	"fmt"
