+++ go/test/endtoend/vtgate/queries/query_fuzzing/schema.sql
