+++ go/test/endtoend/vtgate/queries/query_fuzzing/svschema.json
