+++ go/test/endtoend/vtgate/queries/query_fuzzing/vschema.json
